Move `Request` & `Response` Into their own module
